refactor: club 에서 컨트롤러 - 서비스 간 포트-어댑터 제거
