+++ dongmanee/src/main/java/com/dongmanee/domain/club/controller/ClubController.java
@@ -8,10 +8,10 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.dongmanee.domain.club.controller.apidoc.ClubControllerApiDocs;
 import com.dongmanee.domain.club.controller.mapper.ClubMapper;
-import com.dongmanee.domain.club.controller.port.ClubControllerClubCategoryService;
-import com.dongmanee.domain.club.controller.port.ClubService;
 import com.dongmanee.domain.club.domain.Club;
 import com.dongmanee.domain.club.dto.request.RequestCreateClub;
+import com.dongmanee.domain.club.service.ClubCategoryService;
+import com.dongmanee.domain.club.service.ClubService;
 import com.dongmanee.domain.member.domain.Member;
 import com.dongmanee.domain.member.service.MemberService;
 import com.dongmanee.global.utils.ApiResult;
@@ -29,14 +29,14 @@ public class ClubController implements ClubControllerApiDocs {
 	private final ClubMapper clubMapper;
 	private final ClubService clubService;
 	private final MemberService memberService;
-	private final ClubControllerClubCategoryService clubControllerClubCategoryService;
+	private final ClubCategoryService clubCategoryService;
 
 	//TODO: 클럽 정보 가져오는 URL 리턴으로 변경
 	@PostMapping("/clubs")
 	public ApiResult<?> createClub(@Valid @RequestBody RequestCreateClub createClub,
 		@AuthenticationPrincipal UserDetails userDetails) {
 		Member requestMember = memberService.findById(Long.parseLong(userDetails.getUsername()));
-		Club club = clubMapper.toEntity(createClub, clubControllerClubCategoryService);
+		Club club = clubMapper.toEntity(createClub, clubCategoryService);
 		clubService.createClub(club, requestMember);
 		return ApiResult.isCreated("클럽이 생성되었습니다.");
 	}
