--- dongmanee/src/main/java/com/dongmanee/domain/club/controller/port/ClubControllerClubCategoryService.java
@@ -1,7 +1,0 @@
-package com.dongmanee.domain.club.controller.port;
-
-import com.dongmanee.domain.club.domain.ClubCategory;
-
-public interface ClubControllerClubCategoryService {
-	ClubCategory findById(Long categoryId);
-}
