+++ dongmanee/src/main/java/com/dongmanee/domain/club/service/ClubCategoryService.java
@@ -1,0 +1,7 @@
+package com.dongmanee.domain.club.service;
+
+import com.dongmanee.domain.club.domain.ClubCategory;
+
+public interface ClubCategoryService {
+	ClubCategory findById(Long categoryId);
+}
