+++ dongmanee/src/main/java/com/dongmanee/domain/club/service/ClubCategoryServiceImpl.java
@@ -2,7 +2,6 @@ package com.dongmanee.domain.club.service;
 
 import org.springframework.stereotype.Service;
 
-import com.dongmanee.domain.club.controller.port.ClubControllerClubCategoryService;
 import com.dongmanee.domain.club.domain.ClubCategory;
 import com.dongmanee.domain.club.exception.CategoryNotFoundException;
 import com.dongmanee.domain.club.service.port.ClubCategoryServiceClubCategoryRepository;
@@ -11,7 +10,7 @@ import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
-public class ClubCategoryServiceImpl implements ClubControllerClubCategoryService {
+public class ClubCategoryServiceImpl implements ClubCategoryService {
 	private final ClubCategoryServiceClubCategoryRepository clubCategoryRepository;
 
 	public ClubCategory findById(Long categoryId) {
