+++ dongmanee/src/main/java/com/dongmanee/domain/club/service/ClubService.java
@@ -1,9 +1,12 @@
-package com.dongmanee.domain.club.controller.port;
+package com.dongmanee.domain.club.service;
 
 import com.dongmanee.domain.club.domain.Club;
 import com.dongmanee.domain.club.domain.ClubSns;
+import com.dongmanee.domain.member.domain.Member;
 
-public interface ClubInfoUpdateService {
+public interface ClubService {
+	void createClub(Club club, Member member);
+
 	Club editClubDescriptionAndAddress(Long memberId, Club club);
 
 	ClubSns addClubSns(Long memberId, ClubSns clubSns, Long clubId);
